Update privacy.htm
I assume that the line about storing copies of transcripts contains a typo, and propose a fix here. Should this be intended, then I'd propose to use a positive connotation for the sentence. (e. g. "We do store copies...". Furthermore, <b> doesn't seem to work on your page, probably related to the CSS.